Refactor to have unique like-controller, create route for getting all the liked albums of a specific user
