+++ db/controllers/user-controller.js
@@ -93,34 +93,6 @@ userRouter.get("/lookup/:username", (req, res) => {
         });
 });
 
-// TESTING ROUTE FOR LIKES - remove later
-userRouter.post("/likes", (req, res) => {
-    // Validate request
-    if (!req.body.userId && !req.body.albumId) {
-        res.status(400).send({
-            message: "Fields can not be empty!"
-        });
-        return;
-    }
-    // Create a new like
-    const like = {
-        UserId: req.body.userId,
-        AlbumId: req.body.albumId,
-    };
-
-    // Save User in the database
-    Like.create(like)
-        .then(data => {
-            res.send(data);
-        })
-        .catch(err => {
-            res.status(500).send({
-                message:
-                    err.message || "Some error occurred while creating the Like."
-            });
-        });
-});
-
 const isModerator = async (userId) => {
     try {
       const moderator = await Moderator.findOne({ where: { UserId: userId } });
