+++ server.js
@@ -5,6 +5,7 @@ import UserFollow from "./db/models/userFollow.js";
 import Album from "./db/models/album.js";
 import Track from "./db/models/track.js";
 import Review from "./db/models/review.js";
+import Like from "./db/models/like.js";
 import spotify from "./spotify/index.js";
 import sequelize from "./db/models/index.js";
 import userRouter from "./db/controllers/user-controller.js";
@@ -12,6 +13,7 @@ import albumRouter from "./db/controllers/album-controller.js";
 import trackRouter from "./db/controllers/track-controller.js";
 import reviewRouter from "./db/controllers/review-controller.js";
 import userFollowRouter from "./db/controllers/userFollow-controller.js";
+import likeRouter from "./db/controllers/like-controller.js";
 import dotenv from "dotenv";
 
 dotenv.config();
@@ -36,6 +38,7 @@ app.use('/albums', albumRouter);
 app.use('/tracks', trackRouter);
 app.use('/reviews', reviewRouter);
 app.use('/user-follows', userFollowRouter);
+app.use('/likes', likeRouter);
 app.listen(PORT, () => {
     console.log(`Server is running on port ${PORT}.`);
 });
